# openapi-v4.yaml
openapi: 3.0.3
info:
  title: Unified Food Delivery API - Refined
  description: |
    Fully detailed and corrected OpenAPI 3.0 specification for the Unified Food Delivery backend.
    Includes authentication, users, cart, chat, deliveries, geocoding, file uploads,
    analytics, restaurants, foods, orders, payments, feedback, notifications,
    settings, health-check, route optimization, rate limiting, alias routes,
    and WebSocket event summary.
  version: "4.0.0"
servers:
  - url: /api
    description: Primary API server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /cart:
    get:
      tags: [Cart]
      security: [{ BearerAuth: [] }]
      summary: Get cart items
      responses: { "200": { description: OK } }
    post:
      tags: [Cart]
      security: [{ BearerAuth: [] }]
      summary: Add/increment food in cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                foodId: { type: string }
                quantity: { type: integer }
                excludedIngredients:
                  type: array
                  items:
                    type: string
      responses:
        { "201": { description: Created }, "400": { description: Bad Request } }
  /cart/{id}:
    parameters:
      [{ name: id, in: path, required: true, schema: { type: string } }]
    put:
      tags: [Cart]
      security: [{ BearerAuth: [] }]
      summary: Update cart item quantity or exclusions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity: { type: integer }
                excludedIngredients:
                  type: array
                  items:
                    type: string
      responses: { "200": { description: OK } }
    delete:
      tags: [Cart]
      security: [{ BearerAuth: [] }]
      summary: Remove item by ID
      responses: { "204": { description: Deleted } }
  /cart/clear:
    delete:
      tags: [Cart]
      security: [{ BearerAuth: [] }]
      summary: Clear all items from the cart
      responses: { "204": { description: "Cart cleared" } }

  # Chat
...
  /geocode/route:
    get:
      tags: [Geocoding]
      summary: Route directions
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string }
        - name: to
          in: query
          required: true
          schema: { type: string }
        - name: mode
          in: query
          schema: { type: string, default:car }
      responses: { "200": { description: OK } }
...